updated and manually tested plugins
